% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factormodel.R
\name{compute.ife}
\alias{compute.ife}
\title{compute.ife}
\usage{
compute.ife(formla, xformla, data, idname, tname, treat.t)
}
\arguments{
\item{formla}{y ~ d where y is the outcome and d is the treatment indicator}

\item{xformla}{~x where x includes the covariates in the model}

\item{data}{is the dataset}

\item{treat.period}{the numerical value of the time period when individuals become treated}
}
\value{
IFEobj
}
\description{
Does the heavy-lifting of actually computing treatment effects
}
\keyword{internal}
