% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factormodel.R
\name{compute.ife2}
\alias{compute.ife2}
\title{compute.ife2}
\usage{
compute.ife2(formla, xformla, data, idname, tname, treat.t)
}
\arguments{
\item{formla}{y ~ d where y is the outcome and d is the treatment indicator}

\item{xformla}{~x where x includes the covariates in the model}

\item{data}{is the dataset}
}
\value{
IFEobj
}
\description{
does the heavy-lifting of estimating average treatment effects
 in interactive fixed effects models using GMM
}
\keyword{internal}
