% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pre_test.R
\name{cov_pre_test}
\alias{cov_pre_test}
\title{cov_pre_test}
\usage{
cov_pre_test(
  yname,
  gname,
  tname,
  idname,
  data,
  nife,
  xformla = ~1,
  anticipation = 0
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{nife}{the number of interactive fixed effects to include in the model}

\item{xformla}{Formula for which covariates to include in the model.  Default is ~1.}

\item{anticipation}{Number of periods that treatment is anticipated.  Default
is 0.  This is in ``periods''; e.g., code will work in time periods are
equally spaced but 2 years apart.  In this case, to allow for treatment
anticipation of 2 year (<=> 1 period), set \code{anticipation = 1}.}
}
\value{
\code{ptetools::pte_results} object
}
\description{
Pre-test for time invariant effects of covariates in an
interactive fixed effects model for untreated potential outcomes
}
