% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pre_test_attgt.R
\name{cov_pre_test_attgt}
\alias{cov_pre_test_attgt}
\title{cov_pre_test_attgt}
\usage{
cov_pre_test_attgt(
  gt_data,
  nife = 1,
  xformla = ~1,
  zformla,
  anticipation = 0,
  ret_ife_regs = FALSE,
  ...
)
}
\arguments{
\item{gt_data}{data frame that is local to the specific groups
and times for which we'll be computing a treatment efect
estimate.}

\item{nife}{The number of interactive fixed effects.  Default is 1.}

\item{xformla}{Formula for which covariates to include in the model.  Default is ~1.}

\item{zformla}{Formula for moment conditions to identify interactive fixed effects
parameters.  This must include at least \code{nife} additional covariates relative to
xformla.}

\item{anticipation}{Number of periods that treatment is anticipated.  Default
is 0.  This is in ``periods''; e.g., code will work in time periods are
equally spaced but 2 years apart.  In this case, to allow for treatment
anticipation of 2 year (<=> 1 period), set \code{anticipation = 1}.}

\item{ret_ife_regs}{Whether or not to return the first stage ife regressions; default is FALSE.}

\item{...}{extra arguments}
}
\value{
ptetools::attgt_if object
}
\description{
Computes estimates of ATT(g,t), overall ATT, and dynamic effects
under interactive fixed effects model for untreated potential outcomes
using the approach in Callaway and Karami (2021)
}
