% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pre_test_subset.R
\name{cov_pre_test_subset}
\alias{cov_pre_test_subset}
\title{cov_pre_test_subset}
\usage{
cov_pre_test_subset(data, g, tp, nife, anticipation = 0, ...)
}
\arguments{
\item{data}{the full dataset}

\item{g}{the current group}

\item{tp}{the current time period}

\item{nife}{the number of interactive fixed effects}

\item{anticipation}{number of periods before participating
treatment that treatment can have an effect on outcomes}

\item{...}{extra arguments to get the subset correct}
}
\value{
list that contains correct subset of data, \code{n1}
number of observations
in this subset, and \code{disidx} a vector of the correct ids for this
subset.
}
\description{
A function for obtaining the correct subset in a staggered treatment
adoption setup with panel data for conducting a pre-test that the effect of
covariates does not change over time.
}
