% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ife-package}
\alias{ife-package}
\title{Treatment Effects in Interactive Fixed Effects Models}
\description{
The \code{ife} package contains tools for estimating treatment
effects when untreated potential outcomes are generated by an interactive
fixed effects model.
}
\author{
\strong{Maintainer}: Brantly Callaway \email{brantly.callaway@uga.edu}

Authors:
\itemize{
  \item Sonia Karami \email{sonia.karami88@gmail.com}
  \item Emmanuel Tsyawo \email{emmanuel.tsyawo@um6p.ma}
}

}
\keyword{internal}
