% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factormodel.R
\name{ife}
\alias{ife}
\title{ife}
\usage{
ife(formla, xformla, data, idname, tname, treat.t, boot.iters = 100,
  cores = 1)
}
\arguments{
\item{formla}{y ~ d where y is the outcome and d is the treatment indicator}

\item{xformla}{~x where x includes the covariates in the model}

\item{data}{is the dataset}
}
\value{
IFEobj
}
\description{
compute treatment effects in interactive fixed effects models
}
