% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ife.R, R/influence_func_estimand.R
\name{ife}
\alias{ife}
\alias{ife_constructor}
\alias{influence_func_estimate}
\title{Create a new `influence_func_estimate` object}
\usage{
ife(
  x,
  eif,
  weights = rep(1, length(eif)),
  id = as.character(1:length(eif)),
  critical_value = qnorm(0.975)
)

influence_func_estimate(
  x,
  eif,
  weights = rep(1, length(eif)),
  id = as.character(1:length(eif)),
  critical_value = qnorm(0.975)
)
}
\arguments{
\item{x}{[\code{numeric(1)}]\cr
The individual-level point estimate.}

\item{eif}{[\code{numeric(n)}]\cr
The individual-level influence function.}

\item{weights}{[\code{numeric(n)}]\cr
Optional known survey sampling weights.}

\item{id}{[\code{character(n)}]\cr
Optional cluster identifiers.}

\item{critical_value}{[\code{numeric(1)}]\cr
Optional critical value for constructing confidence interval.}
}
\value{
An 'S7' object of class \code{influence_func_estimate}.
}
\description{
Create a new `influence_func_estimate` object
}
\details{
If known survey weights are provided, the variance estimator is the sample variance of the influence function
multiplied by the survey weights (see DOI: 10.1093/aje/kwu197 for more information). If there is clustering, 
\code{x} and \code{eif} are assumed to be on the individual-level. The individual-level influence function is 
then aggregated to the cluster-level and the variance estimator is the sample variance of the estimated 
cluster-level influence function, scaled by the number of clusters (see DOI: 10.1002/sim.9813 for more information).
}
\examples{
.x <- rnorm(100)
.y <- rnorm(100)
x <- influence_func_estimate(mean(.x), .x - mean(.x))
y <- influence_func_estimate(mean(.x), .y - mean(.y))
x + y
x + 1
1 - y
x / y
x * y
tidy(x)
# Example: Confidence interval for a variance estimate
.z <- rnorm(100, 0, sqrt(4))
ife(mean(.z^2), .z^2 - mean(.z^2)) - ife(mean(.z), .z - mean(.z))^2
}
