% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ife_setup_data.R
\name{ife_setup_pte}
\alias{ife_setup_pte}
\title{ife_setup_pte}
\usage{
ife_setup_pte(
  yname,
  gname,
  tname,
  idname,
  data,
  nife,
  anticipation = 0,
  alp = 0.05,
  biters = 100,
  cl = 1,
  ...
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{nife}{the number of interactive fixed effects to include in the model}

\item{anticipation}{how many periods before the treatment actually takes
place that it can have an effect on outcomes}

\item{alp}{significance level; default is 0.05}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}

\item{...}{extra arguments}
}
\description{
Setup panel treatment effects parameters
}
