% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_func_estimand.R
\name{influence_func_estimand}
\alias{influence_func_estimand}
\title{Create a new `influence_func_estimand` object}
\usage{
influence_func_estimand(
  x,
  eif,
  weights = rep(1, length(eif)),
  id = as.character(1:length(eif))
)
}
\arguments{
\item{x}{[\code{numeric(1)}]\cr
The estimand.}

\item{eif}{[\code{numeric(n)}]\cr
The influence function.}

\item{weights}{[\code{numeric(n)}]\cr
Optional sampling weights.}

\item{id}{[\code{character(n)}]\cr
Optional cluster identifiers.}
}
\value{
An 'S7' object of class \code{influence_func_estimand}.
}
\description{
Create a new `influence_func_estimand` object
}
\examples{
x <- influence_func_estimand(5, runif(10))
y <- influence_func_estimand(5, runif(10))
x + y
x + 1
1 - y
x / y
x * y
}
