% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job_displacement_data}
\alias{job_displacement_data}
\title{Individual Panel Data about Job Displacement}
\format{
A data frame with 18264 rows and 10 columns
\describe{
\item{id}{individual identifier}
\item{displaced}{whether an individual is ever displaced from their job in any time period in the sample}
\item{EDUC}{education category}
\item{first.displaced}{for displaced workers, the year that they were displaced; for non-displaced workers, it is equal to 0}
\item{race}{race}
\item{gender}{gender}
\item{afqt}{armed forces qualification test normalized score}
\item{year}{the year of the observation}
\item{earn}{yearly earnings}
\item{learn}{log of yearly earnings}
}
}
\source{
Callaway and Karami (2021)
}
\usage{
job_displacement_data
}
\description{
The data used in Callaway and Karami (2021)
}
\keyword{datasets}
