% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt.R
\name{linear_trends}
\alias{linear_trends}
\title{linear_trends}
\usage{
linear_trends(
  yname,
  gname,
  tname,
  idname,
  data,
  xformla = ~1,
  anticipation = 0,
  cband = TRUE,
  alp = 0.05,
  boot_type = "multiplier",
  biters = 100,
  cl = 1
)
}
\value{
\code{pte::pte_results} object
}
\description{
Compute treatment effects in model with linear trends for
untreated potential outcomes
}
