% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_subset.R
\name{lt_subset}
\alias{lt_subset}
\title{lt_subset}
\usage{
lt_subset(data, g, tp, anticipation = 0, ...)
}
\arguments{
\item{data}{the full dataset}

\item{g}{the current group}

\item{tp}{the current time period}

\item{...}{extra arguments to get the subset correct}
}
\value{
list that contains correct subset of data, \code{n1}
number of observations
in this subset, and \code{disidx} a vector of the correct ids for this
subset.
}
\description{
A function for obtaining the correct subset in a staggered treatment
adoption setup with panel data using a linear trends model for untreated potential
outcomes.
}
