% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{reset.sim}
\alias{reset.sim}
\title{reset.sim A function that sets "reasonable" values for parameters
with which to simulate data using the \code{gen_ife_data} function}
\usage{
reset.sim()
}
\description{
reset.sim A function that sets "reasonable" values for parameters
with which to simulate data using the \code{gen_ife_data} function
}
\keyword{internal}
