% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_temp.R
\name{reset.sim2}
\alias{reset.sim2}
\title{reset.sim A function that sets "reasonable" values for parameters
with which to simulate data using the \code{gen_ife_data} function}
\usage{
reset.sim2()
}
\description{
reset.sim A function that sets "reasonable" values for parameters
with which to simulate data using the \code{gen_ife_data} function
}
\keyword{internal}
