% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_ife.R
\name{staggered_ife}
\alias{staggered_ife}
\title{staggered_ife}
\usage{
staggered_ife(
  yname,
  gname,
  tname,
  idname,
  data,
  nife,
  xformla = ~1,
  ret_ife_regs = TRUE,
  anticipation = 0,
  cband = TRUE,
  alp = 0.05,
  boot_type = "multiplier",
  biters = 100,
  cl = 1
)
}
\arguments{
\item{nife}{the number of interactive fixed effects to include in the model}
}
\value{
\code{pte::pte_results} object
}
\description{
Compute treatment effects in interactive fixed effects models
with a small number of time periods by exploiting staggered treatment
adoption
}
