% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pre_test.R
\name{summary.cov_pretest_results}
\alias{summary.cov_pretest_results}
\title{summary.cov_pretest_results}
\usage{
\method{summary}{cov_pretest_results}(object, ...)
}
\arguments{
\item{object}{a cov_pretest_results object}

\item{...}{other arguments}
}
\description{
summary.cov_pretest_results
}
