% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ife.R, R/imports.R
\docType{package}
\name{ife}
\alias{ife}
\title{ife}
\usage{
ife(
  yname,
  gname,
  tname,
  idname,
  data,
  nife,
  xformla = ~1,
  zformla,
  ret_ife_regs = TRUE,
  anticipation = 0,
  cband = TRUE,
  alp = 0.05,
  boot_type = "multiplier",
  biters = 100,
  cl = 1
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{nife}{the number of interactive fixed effects to include in the model}

\item{xformla}{Formula for which covariates to include in the model.  Default is ~1.}

\item{zformla}{Formula for moment conditions to identify interactive fixed effects
parameters.  This must include at least \code{nife} additional covariates relative to
xformla.}

\item{ret_ife_regs}{Whether or not to return the first stage ife regressions; default is FALSE.}

\item{anticipation}{Number of periods that treatment is anticipated.  Default
is 0.  This is in ``periods''; e.g., code will work in time periods are
equally spaced but 2 years apart.  In this case, to allow for treatment
anticipation of 2 year (<=> 1 period), set \code{anticipation = 1}.}

\item{cband}{whether or not to compute a uniform (instead of
pointwise) confidence band}

\item{alp}{significance level; default is 0.05}

\item{boot_type}{should be one of "multiplier" (the default) or "empirical".
The multiplier bootstrap is generally much faster, but \code{attgt_fun} needs
to provide an expression for the influence function (which could be challenging
to figure out).  If no influence function is provided, then the \code{pte}
package will use the empirical bootstrap no matter what the value of this
parameter.}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}
}
\value{
\code{pte::pte_results} object
}
\description{
Compute treatment effects in interactive fixed effects models
with a small number of time periods

The \code{ife} package contains tools for estimating treatment
effects when untreated potential outcomes are generated by an interactive
fixed effects model.
}
