% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_attgt.R
\name{lt_attgt}
\alias{lt_attgt}
\title{lt_attgt}
\usage{
lt_attgt(gt_data, xformla = ~1, anticipation = 0, ...)
}
\arguments{
\item{xformla}{Formula for which covariates to include in the model.  Default is ~1.}

\item{anticipation}{Number of periods that treatment is anticipated.  Default
is 0.  This is in ``periods''; e.g., code will work in time periods are
equally spaced but 2 years apart.  In this case, to allow for treatment
anticipation of 2 year (<=> 1 period), set \code{anticipation = 1}.}
}
\value{
\code{attgt_if} object
}
\description{
Computes estimates of ATT(g,t) using a linear trends model
for untreated potential outcomes
}
