% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_ife_setup_data.R
\name{staggered_ife_setup_data2}
\alias{staggered_ife_setup_data2}
\title{staggered_ife_setup_data2}
\usage{
staggered_ife_setup_data2(
  yname,
  gname,
  tname,
  idname,
  data,
  required_pre_periods = 1,
  nife = 1,
  anticipation = 0,
  cband = TRUE,
  alp = 0.05,
  boot_type = boot_type,
  biters = 100,
  cl = 1,
  ...
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{required_pre_periods}{the number of pre-treatment time
periods that are needed}

\item{nife}{the number of interactive fixed effects to include in the model}

\item{anticipation}{Number of periods that treatment is anticipated.  Default
is 0.  This is in ``periods''; e.g., code will work in time periods are
equally spaced but 2 years apart.  In this case, to allow for treatment
anticipation of 2 year (<=> 1 period), set \code{anticipation = 1}.}

\item{cband}{whether to compute a uniform (instead of pointwise)
confidence band}

\item{alp}{significance level; default is 0.05}

\item{boot_type}{type of bootstrap}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}

\item{...}{extra arguments}
}
\value{
\code{pte_params} object
}
\description{
This function sets up the periods/groups for using
\code{staggered_ife_attgt2} (the main approach proposed in Callaway and
Tsyawo (2023)).
}
