% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered_ife_setup_data.R
\name{staggered_ife_setup_pte}
\alias{staggered_ife_setup_pte}
\title{staggered_ife_setup_data}
\usage{
staggered_ife_setup_pte(
  yname,
  gname,
  tname,
  idname,
  data,
  required_pre_periods = 1,
  nife = 1,
  anticipation = 0,
  cband = TRUE,
  alp = 0.05,
  boot_type = boot_type,
  biters = 100,
  cl = 1,
  ...
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{alp}{significance level; default is 0.05}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}
}
\value{
\code{pte_params} object
}
\description{
Estimate treatment effects in an interactive fixed effects
model for untreated potential outcomes by exploiting staggered treatment
adoption.
}
